move identifier fixtures to share lib file and refactor tests to remove duplicate fixture declarations
